+++ widget_driver/example/lib/pub_dev_example_code/my_first_drivable_widget_driver.g.dart
@@ -8,7 +8,7 @@ part of 'my_first_drivable_widget_driver.dart';
 
 // coverage:ignore-file
 
-// This file was generated with widget_driver_generator version "1.2.0"
+// This file was generated with widget_driver_generator version "1.3.6"
 
 class _$TestMyFirstDrivableWidgetDriver extends TestDriver implements MyFirstDrivableWidgetDriver {
   @override