+++ widget_driver/example/lib/widgets/coffee_community/coffee_library/coffee_row.dart
@@ -31,7 +31,7 @@ class CoffeeRow extends StatelessWidget {
           borderRadius: const BorderRadius.all(Radius.circular(8)),
           boxShadow: [
             BoxShadow(
-              color: Colors.grey.withOpacity(0.5),
+              color: Colors.grey.withValues(alpha: 0.5),
               spreadRadius: 1,
               blurRadius: 1,
               offset: const Offset(0, 2), // changes position of shadow