+++ widget_driver/example/lib/widgets/home_page/home_page_driver.g.dart
@@ -8,7 +8,7 @@ part of 'home_page_driver.dart';
 
 // coverage:ignore-file
 
-// This file was generated with widget_driver_generator version "1.2.0"
+// This file was generated with widget_driver_generator version "1.3.6"
 
 class _$TestHomePageDriver extends TestDriver implements HomePageDriver {
   @override