+++ widget_driver/example/lib/widgets/playground/playground_test_widget/playground_test_child_widget_driver.g.dart
@@ -8,7 +8,7 @@ part of 'playground_test_child_widget_driver.dart';
 
 // coverage:ignore-file
 
-// This file was generated with widget_driver_generator version "1.2.0"
+// This file was generated with widget_driver_generator version "1.3.6"
 
 class _$TestPlaygroundTestChildWidgetDriver extends TestDriver implements PlaygroundTestChildWidgetDriver {
   @override